Ensure spawned entities are separate instances
The issue is that Gson only creates a single instance of each monster
during JSON parsing, so multiple of the same type end up being the exact
same instance. Right now this is overcome by creating a new monsterList
from JSON data for each spawn event, which is probably more cumbersome
than it needs to be.